Refactor: Split four Gerrit modules
models, comparator and urls become packages; client stays a module and
sheds two siblings instead.

The difference matters.  Tests monkeypatch GerritRestAPI, HTTPBasicAuth
and get_credentials_for_host in gerrit.client's own namespace, fifty
sites in total.  Had client become a package with those call sites in a
submodule, every one of those patches would still have applied cleanly
to the package attribute while the code under test resolved the real
name — passing for the wrong reason, or reaching the network.  So the
functions that reference a patched name stay in client.py, and only
code that does not touch them moved to _client_errors.py and
_client_requests.py.

That reasoning was checked rather than assumed: breaking the relocated
_request made test_get_normalizes_path fail, and making __init__ bypass
the module global made two tests fail, one of them with a live DNS
lookup.  Both confirm the patches still reach the code they cover.

comparator and client each needed a mixin to move method bodies off an
oversized class without renaming anything or altering the method
surface.  No method body or call site changed.

discover_base_path loses three helpers to come under the length limit:
one for the repeated cache-and-log block, one wrapping a single probe,
and one driving the probe loop and its time budget.

The Gerrit URL hygiene allow-lists follow urls.py to urls/builder.py.
discovery.py needs no entry: it builds only probe URLs, which no
pattern matches.

Findings drop from 47 to 42; the score rises from 44 to 48.
